IGNITE-17167 ConfigurationAsmGenerator decomposition. Part 1, InnerNode-related classes.
